requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-extensions==3.2.1
idna==3.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
Werkzeug==2.2.3
